+++ R/PipeOpClassBalancing.R
@@ -109,7 +109,7 @@ PipeOpClassBalancing = R6Class("PipeOpClassBalancing",
           levels = c("all", "major", "minor", "nonmajor", "nonminor", "one"), tags = "train"),
         ParamFct$new("adjust",
           levels = c("all", "major", "minor", "nonmajor", "nonminor", "upsample", "downsample"), tags = "train"),
-        ParamLgl$new("shuffle", default = TRUE, tags = "train")
+        ParamLgl$new("shuffle", tags = "train")
       ))
       ps$values = list(ratio = 1, reference = "all", adjust = "all", shuffle = TRUE)
       super$initialize(id, param_set = ps, param_vals = param_vals, can_subset_cols = FALSE, task_type = "TaskClassif", tags = "imbalanced data")