+++ R/zzz.R
@@ -13,9 +13,8 @@ register_mlr3 = function() {
   x = utils::getFromNamespace("mlr_reflections", ns = "mlr3")
   x$pipeops$valid_tags = unique(c(x$pipeops$valid_tags,
     c("abstract", "meta", "missings", "feature selection", "imbalanced data",
-    "data transform", "target transform", "ensemble", "robustify", "learner", "encode", "multiplicity")))
-   x$learner_predict_types$classif$impact = "impact"
-   x$learner_predict_types$regr$impact = "impact"
+    "data transform", "target transform", "ensemble", "robustify", "learner", "encode",
+     "multiplicity")))
 }
 
 .onLoad = function(libname, pkgname) {  # nocov start
@@ -35,3 +34,8 @@ register_mlr3 = function() {
    pkgname = vapply(hooks[-1], function(x) environment(x)$pkgname, NA_character_)
    setHook(event, hooks[pkgname != "mlr3pipelines"], action = "replace")
 } # nocov end
+
+# static code checks should not complain about commonly used data.table columns
+utils::globalVariables(c("src_id", "dst_id", "name", "op.id", "response", "truth"))
+
+leanify_package()