+++ tests/testthat/test_learner_weightedaverage.R
@@ -1,6 +1,7 @@
 context("WeightedAverage Learner")
 
 test_that("LearnerClassifAvg", {
+  skip_on_cran()  # takes too long
   lrn = LearnerClassifAvg$new()
   expect_learner(lrn)
   df = data.frame(x = matrix(sample(c("a", "b", "c"), 100, replace = TRUE), nrow = 10), y = as.factor(sample(c("a", "b", "c"), 10, replace = TRUE)), stringsAsFactors = TRUE)
@@ -16,7 +17,7 @@ test_that("LearnerClassifAvg", {
   expect_prediction(prd)
 
 
-  df = data.frame(x = matrix(runif(100), nrow = 10), y = as.factor(sample(c(0, 1), 10, replace = TRUE)))
+  df = data.frame(x = matrix(runif(90), nrow = 30), y = as.factor(sample(c(0, 1), 30, replace = TRUE)))
   tsk = TaskClassif$new(id = "tsk", backend = df, target = "y")
 
   for (predicttype in c("prob", "response")) {
@@ -26,7 +27,7 @@ test_that("LearnerClassifAvg", {
     lrn = LearnerClassifAvg$new()
     lrn$predict_type = predicttype
     expect_learner(lrn)
-    lrn$param_set$values = list(measure = msr("classif.acc"), algorithm = "NLOPT_LN_COBYLA")
+    lrn$param_set$values$measure = msr("classif.acc")
     lrn$train(intask)
     expect_list(lrn$model, names = "named")
     expect_numeric(lrn$model$weights, len = 3)
@@ -41,19 +42,18 @@ test_that("LearnerClassifAvg", {
       lrn = LearnerClassifAvg$new()
       lrn$predict_type = predicttype
       expect_learner(lrn)
-      lrn$param_set$values = list(measure = msr("classif.auc"), algorithm = "NLOPT_LN_COBYLA")
+      lrn$param_set$values$measure = msr("classif.auc")
       lrn$train(intask)
       expect_list(lrn$model, names = "named")
       expect_numeric(lrn$model$weights, len = 3)
       prd = lrn$predict(intask)
       expect_prediction(prd)
     }
-    skip_on_cran()  # skip on cran after one round
   }
 })
 
 test_that("LearnerRegrAvg", {
-  skip_on_cran()  # skip on cran after one round
+  skip_on_cran()  # takes too long
   lrn = LearnerRegrAvg$new()
   expect_learner(lrn)
   df = data.frame(x = matrix(rnorm(100), nrow = 10), y = rnorm(100))
@@ -90,7 +90,7 @@ test_that("LearnerRegrAvg", {
   lrn = LearnerRegrAvg$new()
   lrn$predict_type = "response"
   expect_learner(lrn)
-  lrn$param_set$values = list(measure = msr("regr.mse"), algorithm = "NLOPT_LN_COBYLA")
+  lrn$param_set$values$measure = msr("regr.mse")
   lrn$train(intask)
   expect_list(lrn$model, names = "named")
   expect_numeric(lrn$model$weights, len = 3)
@@ -100,7 +100,7 @@ test_that("LearnerRegrAvg", {
 })
 
 test_that("LearnerClassifAvg Pipeline", {
-  skip_on_cran()  # skip on cran after one round
+  skip_on_cran()  # takes too long
   tsk = mlr_tasks$get("iris")
   # Works for response
   # TODO: this is a bit of a deep problem: https://github.com/mlr-org/mlr3pipelines/issues/216
@@ -146,7 +146,7 @@ test_that("LearnerClassifAvg Pipeline", {
 })
 
 test_that("LearnerRegrAvg Pipeline", {
-  skip_on_cran()  # skip on cran after one round
+  skip_on_cran()  # takes too long
   tsk = mlr_tasks$get("boston_housing")
   # Works for response
   # TODO: this is a bit of a deep problem: https://github.com/mlr-org/mlr3pipelines/issues/216