+++ tests/testthat/test_mlr_graphs_bagging.R
@@ -2,6 +2,7 @@ context("ppl - pipeline_bagging")
 
 
 test_that("Bagging Pipeline", {
+  skip_on_cran()  # takes too long
   # classif
   tsk = tsk("iris")
   lrn = lrn("classif.rpart")