+++ tests/testthat/test_pipeop_ovr.R
@@ -22,7 +22,7 @@ test_that("PipeOpOVRSplit - train and predict", {
   expect_named(tout[[1]], tsk$class_names)
   expect_true(all(pmap_lgl(list(tout[[1]], names(tout[[1]])), .f = function(task, name) {
     expect_task(task)
-    task$target_names == tsk$target_names && task$positive == name && task$negative == "rest." &&
+      all(task$target_names == tsk$target_names) && task$positive == name && task$negative == "rest." &&
       all.equal(task$truth(), factor(ifelse(tsk$truth() == task$positive, task$positive, "rest."), levels = c(task$positive, "rest.")))
   })))
 
@@ -124,4 +124,12 @@ test_that("PipeOpOVRSplit and PipeOpOVRUnite - train and predict", {
   # setting weights to zero results in uniform probs
   gr$param_set$values$ovrunite.weights = rep(0, 3)
   expect_true(all.equal(unique(gr$predict(tsk0)[[1]]$prob), t(c(a = 1/3, b = 1/3, c = 1/3))))
+})
+
+test_that("PipeOpOVRSplit and PipeOpOVRUnite - task size", {
+  gr = PipeOpOVRSplit$new() %>>% LearnerClassifRpart$new() %>>% PipeOpOVRUnite$new()
+  gr$train(tsk("iris")$filter(c(1:30, 51:80, 101:130)))
+  prd = gr$predict(tsk("iris")$filter(c(1:30, 51:80, 101:130)))[[1]]
+  expect_prediction_classif(prd)
+  expect_true(nrow(prd$data$prob) == 90)
 })